pyproject.toml
[tool.poetry]
name = "digital-assistant-first"
version = "1"
description = ""
authors = ["ZakatZakat <comradefrunze@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
streamlit = ">=1.41.1,<2.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
langchain-openai = ">=0.3.3,<0.4.0"
pymupdf = ">=1.25.2,<2.0.0"
streamlit-aggrid = ">=1.1.0,<2.0.0"
langchain-community = ">=0.3.16,<0.4.0"
google-search-results = ">=2.4.2,<3.0.0"
apscheduler = ">=3.11.0,<4.0.0"
faiss-cpu = "<=1.9.0"
openpyxl = ">=3.1.5,<4.0.0"
pydantic-ai = {extras = ["logfire"], version = ">=0.0.21,<0.0.22"}
chromadb = "^0.6.3"
langchain-chroma = "^0.2.1"
pysqlite3-binary = "^0.5.4"
telethon = ">=1.38.1,<2.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


